improve sizing of columns
Before, only the adjacent column was scaled to make space for a new
column width.
Now, the adjacent column is resized until it has reached the min column
width. Afterwards, the next column is used to make space for a resized
column. This effect continues till no column is left that can be
resized. Before a column to the left is touched, all columns to the
right of the primarily have to reach the min column width limit.