fix columns content type
It is not guarantied that `showOrder` is an increasing chain of numbers
without gaps. The user can specify `showOrder`manually, resulting in a
`showOrder` of 7 for example even though it technically is the second
child of the columns content and therefore should get the width of the
second column assigned.